feat:최근 리뷰 디자인 수정
* AlbumReviewSummary 추가
